dx(hub): rename
